+++ resources/views/Templates/newSidebar.blade.php
@@ -76,7 +76,7 @@
                                     <ul class="menu-dropdown">
                                         <li id="menuJobOrder"><a class="waves-effect" href="/administrator/transactions/joborders"><i class="ion ion-ios-list"></i>Job Order</a></li>
                                         <li id="menuTugboatAssignment"><a class="waves-effect" href="/administrator/transactions/tugboatassignment"><i class="ion ion-android-boat"></i>Tugboat Assignment</a></li>
-                                        <li id="menuTeamBuilder"><a class="waves-effect" href="/administrator/transactions/teamassignment"><i class="ion ion-ios-people"></i>Team Builder</a></li>
+                                        <li id="menuTeamBuilder"><a class="waves-effect" href="/administrator/transactions/dispatchandhauling/teamassignment"><i class="ion ion-ios-people"></i>Team Builder</a></li>
                                         <li id="menuScheduling"><a href="/administrator/transactions/scheduling"><i class="fas fa-calendar-alt"></i>Schedule</a></li>
                                         <li id="menuHauling"><a href="/administrator/transactions/hauling"><i class="fas fa-link"></i>Hauling</a></li>
                                         
@@ -99,11 +99,11 @@
                                 
                                 <li><a href="#" class="has-dropdown waves-effect"><i class="ion ion-android-boat"></i>Dispatch &amp; Hauling</a>
                                     <ul class="menu-dropdown">
-                                        <li id="menuJobOrder"><a class="waves-effect" href="/administrator/transactions/joborders"><i class="ion ion-ios-list"></i>Job Order</a></li>
-                                        <li id="menuTugboatAssignment"><a class="waves-effect" href="/administrator/transactions/tugboatassignment"><i class="ion ion-android-boat"></i>Tugboat Assignment</a></li>
-                                        <li id="menuTeamBuilder"><a class="waves-effect" href="/administrator/transactions/teamassignment"><i class="ion ion-ios-people"></i>Team Builder</a></li>
-                                        <li id="menuScheduling"><a href="/administrator/transactions/scheduling"><i class="fas fa-calendar-alt"></i>Schedule</a></li>
-                                        <li id="menuHauling"><a href="/administrator/transactions/hauling"><i class="fas fa-link"></i>Hauling</a></li>
+                                        <li id="menuJobOrder"><a class="waves-effect" href="/affiliates/transactions/joborders"><i class="ion ion-ios-list"></i>Job Order</a></li>
+                                        <li id="menuTugboatAssignment"><a class="waves-effect" href="/affiliates/transactions/tugboatassignment"><i class="ion ion-android-boat"></i>Tugboat Assignment</a></li>
+                                        <li id="menuTeamBuilder"><a class="waves-effect" href="/affiliates/transactions/teamassignment"><i class="ion ion-ios-people"></i>Team Builder</a></li>
+                                        <li id="menuScheduling"><a href="/affiliates/transactions/scheduling"><i class="fas fa-calendar-alt"></i>Schedule</a></li>
+                                        <li id="menuHauling"><a href="/affiliates/transactions/hauling"><i class="fas fa-link"></i>Hauling</a></li>
                                         {{-- <li id="menuScheduling"><a class="waves-effect" href=""><i class="ion ion-ios-time"></i>Scheduling</a></li> --}}
                                     </ul>
                                 </li>
@@ -133,6 +133,7 @@
                             <li><a href="/contracts"><i class="ion ion-ios-paper"></i> Contracts</a></li>
                         </ul>
                     </li>
+                    {{-- if --}}
                     <li>
                         <a id="logout" onclick="event.preventDefault(); document.getElementById('logout-form-admin').submit();"href="/administrator/logout"><i class="ion ion-log-out"></i><span>Logout</span></a>
                     </li>