+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 ### `Pre Final Repository Monday August 27,2018`
 
 &nbsp;
+#.0.0.9 - 2018-09-13 α
+
+`new`
+- [x] requestTeamModal
+- [x] requestTugboatModal
+- [x] Request Team
+- [x] Request Tugboat
+- [x] Summary of Selected Employees in Team Builder
+- [x] Remove Team From Tugboat
+- [x] Remove Employees From Team
+
+`fixes`
+- [x] More Info modal in Job Orders Properly Shows
+- [x] Forward Job Order Now Works
 
 # 0.0.8 - 2018-09-06 α